requirements.txt
datasets>=2.14.0
flask>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
requests>=2.31.0
huggingface_hub>=0.20.0
